+++ src/models/User.php
@@ -58,6 +58,7 @@ class User extends ActiveRecord
     use StatusAttributeTrait;
 
     public const STATUS_ENABLED = 1;
+    public const STATUS_DEFAULT = self::STATUS_ENABLED;
 
     public const AUTH_USER_CREATE = 'userCreate';
     public const AUTH_USER_DELETE = 'userDelete';
@@ -178,7 +179,7 @@ class User extends ActiveRecord
                 'unique',
                 'message' => Yii::t('skeleton', 'This username is already used by another user.'),
                 'skipOnError' => true,
-                'when' => fn () => $this->isAttributeChanged('name')
+                'when' => fn() => $this->isAttributeChanged('name')
             ],
             [
                 ['email'],
@@ -195,7 +196,7 @@ class User extends ActiveRecord
                 'unique',
                 'message' => Yii::t('skeleton', 'This email is already used by another user.'),
                 'skipOnError' => true,
-                'when' => fn () => $this->isAttributeChanged('email')
+                'when' => fn() => $this->isAttributeChanged('email')
             ],
             [
                 ['city', 'first_name', 'last_name'],